GoITeens_UA_PYTHON_1y_3_22_02_24/50_internet_shop/internet_shop-3_frontend/pyproject.toml
[tool.poetry]
name = "internet-shop-3-frontend"
version = "0.1.0"
description = ""
authors = ["Oleksandr Kondratiuk <alex_kondr@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.1.0"
requests = "^2.32.3"
flask-wtf = "^1.2.2"
python-dotenv = "^1.0.1"
email-validator = "^2.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
